pyproject.toml
[project]
name = "civiscope"
version = "0.1.0"
description = "Open source platform for municipal public data ingestion and contract analysis."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0,<6.0",
    "psycopg[binary]>=3.1",
    "httpx>=0.27",
    "pydantic>=2.6",
    "tenacity>=8.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.0",
]
